Move certificates controller policies under `/internal/controller`
