+++ internal/BUILD.bazel
@@ -16,8 +16,8 @@ filegroup(
         "//internal/apis/config/webhook:all-srcs",
         "//internal/apis/meta:all-srcs",
         "//internal/controller/feature:all-srcs",
+        "//internal/controller/policies:all-srcs",
         "//internal/ingress:all-srcs",
-        "//internal/policies:all-srcs",
         "//internal/test/paths:all-srcs",
         "//internal/vault:all-srcs",
         "//internal/webhook/feature:all-srcs",
