+++ internal/controller/policies/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
         "gatherer.go",
         "policies.go",
     ],
-    importpath = "github.com/jetstack/cert-manager/internal/policies",
+    importpath = "github.com/jetstack/cert-manager/internal/controller/policies",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/apis/certmanager/v1:go_default_library",
