+++ pkg/controller/certificates/issuing/internal/BUILD.bazel
@@ -9,11 +9,13 @@ go_library(
     importpath = "github.com/jetstack/cert-manager/pkg/controller/certificates/issuing/internal",
     visibility = ["//pkg/controller/certificates:__subpackages__"],
     deps = [
+        "//internal/controller/feature:go_default_library",
         "//pkg/api/util:go_default_library",
         "//pkg/apis/certmanager/v1:go_default_library",
         "//pkg/apis/meta/v1:go_default_library",
         "//pkg/logs:go_default_library",
         "//pkg/util:go_default_library",
+        "//pkg/util/feature:go_default_library",
         "//pkg/util/pki:go_default_library",
         "@com_github_pavel_v_chernykh_keystore_go_v4//:go_default_library",
         "@com_sslmate_software_src_go_pkcs12//:go_default_library",
@@ -41,6 +43,7 @@ go_test(
         "//pkg/apis/meta/v1:go_default_library",
         "//pkg/controller:go_default_library",
         "//pkg/controller/test:go_default_library",
+        "//pkg/util/feature:go_default_library",
         "//pkg/util/pki:go_default_library",
         "//test/unit/coreclients:go_default_library",
         "//test/unit/crypto:go_default_library",
