+++ pkg/controller/certificates/readiness/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "github.com/jetstack/cert-manager/pkg/controller/certificates/readiness",
     visibility = ["//visibility:public"],
     deps = [
-        "//internal/policies:go_default_library",
+        "//internal/controller/policies:go_default_library",
         "//pkg/api/util:go_default_library",
         "//pkg/apis/certmanager/v1:go_default_library",
         "//pkg/apis/meta/v1:go_default_library",
@@ -50,7 +50,7 @@ go_test(
     srcs = ["readiness_controller_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//internal/policies:go_default_library",
+        "//internal/controller/policies:go_default_library",
         "//pkg/apis/certmanager/v1:go_default_library",
         "//pkg/apis/meta/v1:go_default_library",
         "//pkg/controller:go_default_library",
