+++ pkg/controller/certificates/trigger/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "github.com/jetstack/cert-manager/pkg/controller/certificates/trigger",
     visibility = ["//visibility:public"],
     deps = [
-        "//internal/policies:go_default_library",
+        "//internal/controller/policies:go_default_library",
         "//pkg/api/util:go_default_library",
         "//pkg/apis/certmanager/v1:go_default_library",
         "//pkg/apis/meta/v1:go_default_library",
@@ -51,7 +51,7 @@ go_test(
     srcs = ["trigger_controller_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//internal/policies:go_default_library",
+        "//internal/controller/policies:go_default_library",
         "//pkg/apis/certmanager/v1:go_default_library",
         "//pkg/controller:go_default_library",
         "//pkg/controller/test:go_default_library",
